Split gs_patterns.cpp into a few headers and source files in prep for library separation.
